HPA: fix panic on pre-v2 partially-supplied rules
The description in the autoscaling API for the HorizontalPodAutoscaler
suggests that HorizontalPodAutoscalerSpec's Behavior field (and its
ScaleUp and ScaleDown members) are optional. And that if not supplied,
defaults will be used.

That's true if the entire Behavior is nil because, we go
through "normalizeDesiredReplicas" instead of
"normalizeDesiredReplicasWithBehaviors", but if the structure is only
partially supplied, leaving some members nil, it results in nil
dereferences when we end up going though
normalizeDesiredReplicasWithBehaviors.

So we end up in a situation where:
- If Behavior is entirely absent (nil) we use defaults
- If Behavior is partially specified we panic
- If ScaleUp and ScaleDown are specified, but
stabilizationWindowSeconds is nil in either of them, we panic

In general, this only happens with pre-v2 HPA objects because v2 does
properly fill in the default values.

This can be fixed by using the defaulters to ensure that unpopulated
fields in the behavior objects get filled in with their defaults, so
they can safely be dereferenced by later code and defaults actually
get used as intended.