Wait for resource quota status to be populated
Trying to create a pod using a resourcequota before the status is
populated results in API admission returing a 403 Forbidden. The pod
resize tests for InPlacePodVerticalScaling do not wait for this, and as
a result if the resourcequota controller is too slow adding the status
during say, a busy test run, this test will fail because it makes the
pod before it has the resourcequota status.

This just makes the test wait up to a minute for the status gets
populated. The status generation timing/logic gets tested separately and
it waits a minute, so we wait a minute. I know it's getting populated
because those other tests are passing, this test here just doesn't wait.